pyproject.toml
[project]
name = "lecturaserver"
version = "0.1.0"
description = ""
authors = [{ name = "Anthony Sung", email = "sungpinyue@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-unfold (>=0.76.0,<0.77.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt[crypto] (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-rest-passwordreset (>=1.5.0,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "weasyprint (>=67.0,<68.0)",
    "markdown (>=3.10,<4.0)",
    "pygments (>=2.19.2,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
exclude = ["migrations", ".git", ".nox", ".venv", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"config/settings.py" = ["E402", "F403", "F401"]

[dependency-groups]
dev = ["ruff (>=0.14.13,<0.15.0)"]
